pkg/crypto: move the crypto utils from pkg/repository to pkg
